autoscaler/requirements.txt
# Core dependencies
requests==2.31.0
prometheus-client==0.17.1
kubernetes==28.1.0
PyYAML==6.0.1
urllib3==1.26.18
docker==6.1.3

# API dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database dependencies
pymongo==4.6.0
redis==5.0.1
python-dateutil==2.9.0.post0

# Environment variables
python-dotenv==1.0.0

# Note: gosu is a system package, not a Python package